requirements.txt
requests==2.23.0
Pillow==7.1.2
pathlib==1.0.1
instabot==0.117.0
python-dotenv==0.13.0